refactor: 서비스 인터페이스 분리 및 구현체 분리 작업
