+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -16,6 +16,7 @@ import roomescape.application.dto.ReservationTimeAvailableResponse;
 import roomescape.application.dto.ReservationTimeRequest;
 import roomescape.application.dto.ReservationTimeResponse;
 import roomescape.application.service.ReservationTimeService;
+import roomescape.application.service.ReservationTimeServiceImpl;
 
 @RestController
 @RequestMapping("times")
@@ -23,7 +24,7 @@ public class ReservationTimeController {
 
     private final ReservationTimeService reservationTimeService;
 
-    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+    public ReservationTimeController(ReservationTimeServiceImpl reservationTimeService) {
         this.reservationTimeService = reservationTimeService;
     }
