+++ src/test/java/roomescape/application/service/ReservationServiceImplTest.java
@@ -19,12 +19,13 @@ import roomescape.dao.ReservationTimeDao;
 import roomescape.dao.ThemeDao;
 import roomescape.domain.ReservationTime;
 
-class ReservationServiceTest {
+class ReservationServiceImplTest {
 
     private final ReservationDao reservationDao = new FakeReservationDao();
     private final ReservationTimeDao timeDao = new FakeReservationTimeDao();
     private final ThemeDao themeDao = new FakeThemeDao();
-    private final ReservationService reservationService = new ReservationService(reservationDao, timeDao, themeDao);
+    private final ReservationServiceImpl reservationService = new ReservationServiceImpl(reservationDao, timeDao,
+            themeDao);
 
     @Test
     @Disabled
